resolving bulk upload

+++ Application/DTOs/TelemetryResponseDto.cs
@@ -47,11 +47,11 @@ namespace Application.DTOs
     public class TelemetryStats
     {
         public int Count { get; set; }
-        public double Min { get; set; }
-        public double Max { get; set; }
-        public double Average { get; set; }
-        public double FirstValue { get; set; }
-        public double LastValue { get; set; }
+        public float Min { get; set; }
+        public float Max { get; set; }
+        public float Average { get; set; }
+        public float FirstValue { get; set; }
+        public float LastValue { get; set; }
         public DateTime FirstTimestamp { get; set; }
         public DateTime LastTimestamp { get; set; }
     }
@@ -59,6 +59,6 @@ namespace Application.DTOs
     public class TelemetryPointDto
     {
         public DateTime Time { get; set; }
-        public double Value { get; set; }
+        public float Value { get; set; }
     }
 }
